back-end/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.3.4
martor==1.6.18
mccabe==0.7.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
webencodings==0.5.1
zipp==3.13.0
